misc: migrate project from SDL2 to SDL3

+++ src/game.h
@@ -2,11 +2,10 @@
 
 #pragma once
 
-#include <SDL_render.h>
-#include <SDL_video.h>
 #include <memory>
 
-#include "SDL_stdinc.h"
+#include "SDL3/SDL_render.h"
+#include "SDL3/SDL_video.h"
 #include "scene.h"
 
 namespace spaceshooter {
@@ -37,7 +36,7 @@ class Game {
  private:
   Game(int width, int height);
   void HandleEvent();
-  void Update(Uint32 delta_time);
+  void Update(Uint64 delta_time);
   void Render();
 
   bool is_running_;
